Update EDA/data pipeline diagram
Removed the save dataset step in the end because we will be saving checkpoints throughout this pipeline. Adding one in the end will look like we're only doing it for the last one
Created a separate section for class rebalancing because it is a work right before data test split, and it doesn't really fit into EDA and feature engineering
Few wording changes.